+++ pywps/processing/basic.py
@@ -4,47 +4,7 @@
 # licensed under MIT, Please consult LICENSE.txt for details     #
 ##################################################################
 
-import tempfile
-import pywps.configuration as config
-
-
-class Job(object):
-    def __init__(self, process, wps_request, wps_response):
-        self.process = process
-        self.method = '_run_process'
-        self.wps_request = wps_request
-        self.wps_response = wps_response
-
-    @property
-    def name(self):
-        return self.process.identifier
-
-    @property
-    def workdir(self):
-        return self.process.workdir
-
-    @property
-    def uuid(self):
-        return self.process.uuid
-
-    def dump(self):
-        import dill
-        filename = tempfile.mkstemp(prefix='job_', suffix='.dump', dir=self.workdir)[1]
-        with open(filename, 'w') as fp:
-            dill.dump(self, fp)
-            return filename
-        return None
-
-    @classmethod
-    def load(cls, filename):
-        import dill
-        with open(filename) as fp:
-            job = dill.load(fp)
-            return job
-        return None
-
-    def run(self):
-        getattr(self.process, self.method)(self.wps_request, self.wps_response)
+from pywps.processing.job import Job
 
 
 class Processing(object):
