+++ setup.py
@@ -61,7 +61,7 @@ CONFIG = {
     'scripts': [],
     'entry_points': {
         'console_scripts': [
-            'launch=pywps.processing.joblauncher:main', ]},
+            'joblauncher=pywps.processing.job:launcher', ]},
 }
 
 if sys.version_info.major < 3:
